Move integrators to separate sub-package
