+++ structuralcodes/sections/section_integrators/_section_integrator.py
@@ -1,0 +1,47 @@
+"""Abstract base class for section integrators."""
+
+import abc
+
+from numpy.typing import ArrayLike
+
+from structuralcodes.geometry import CompoundGeometry
+
+
+class SectionIntegrator(abc.ABC):
+    """Abstract base class for section integrators."""
+
+    @abc.abstractmethod
+    def prepare_input(self, geo: CompoundGeometry, strain: ArrayLike):
+        """Prepare general input to the integration method.
+
+        Args:
+            geo (CompoundGeometry): The geometry to integrate over.
+            strain (ArrayLike): The scalar strain components necessary for
+                describing the assumed strain distribution over the geometry.
+        """
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def integrate(self, *prepared_input, **kwargs) -> float:
+        """Integrate stresses over the gemetry.
+
+        The input to this method is generated by the prepare_input method. It
+        also takes kwargs depending on the concrete implementation.
+
+        Arguments:
+            *prepared_input: The input prepared by the prepare_input method.
+
+        Keyword Arguments:
+            **kwargs: Keyword arguments depending on the concrete
+                implementation.
+        """
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def integrate_strain_response_on_geometry(
+        self, geo: CompoundGeometry, strain: ArrayLike, **kwargs
+    ):
+        """Integrate over the geometry to obtain the response due
+        to strains.
+        """
+        raise NotImplementedError
